ZFS compatibility.

+++ install/node.sh
@@ -4,21 +4,25 @@ cd "$(dirname "$0")"
 
 source libfns.sh
 
-while getopts “d:H:S:h” OPTION; do
+while getopts "d:f:H:S:h" OPTION; do
     case $OPTION in
         h)
-            echo "usage: $0 -H [node-host] -S [shipbuilder-host] [ACTION]" 1>&2
+            echo "usage: $0 -H [node-host] -S [shipbuilder-host] -d [device] -f [lxc-filesystem] [ACTION]" 1>&2
             echo '' 1>&2
             echo 'This is the node installer.' 1>&2
             echo '' 1>&2
             echo '  ACTION                Action to perform. Available actions are: list-devices, install'
-            echo '  -H [node-host]        Node hostname' 1>&2
+            echo '  -S [shipbuilder-host] ShipBuilder server user@hostname (flag can be omitted if auto-detected from env/SB_SSH_HOST)' 1>&2
+            echo '  -H [node-host]        Node user@hostname' 1>&2
             echo '  -d [device]           Device to install filesystem on' 1>&2
-            echo '  -S [shipbuilder-host] ShipBuilder server hostname' 1>&2
+            echo '  -f [lxc-filesystem]   LXC filesystem to use; "zfs" or "btrfs" (flag can be ommitted if auto-detected from env/LXC_FS)' 1>&2
             exit 1
             ;;
         d)
             device=$OPTARG
+            ;;
+        f)
+            lxcFs=$OPTARG
             ;;
         H)
             nodeHost=$OPTARG
@@ -35,11 +39,16 @@ shift $(($OPTIND - 1))
 action=$1
 
 test -z "${sbHost}" && autoDetectServer
+test -z "${lxcFs}" && autoDetectFilesystem
 
 # Validate required parameters.
 test -z "${sbHost}" && echo 'error: missing required parameter: -S [shipbuilder-host]' 1>&2 && exit 1
 test -z "${nodeHost}" && echo 'error: missing required parameter: -H [node-host]' 1>&2 && exit 1
-test -z "${action}" && echo 'error: missing required parameter: action' 1>&2 && exit 1
+#test -z "${action}" && echo 'error: missing required parameter: action' 1>&2 && exit 1
+if test -z "${action}"; then
+    echo 'info: action defaulted to: install'
+    action='install'
+fi
 
 
 verifySshAndSudoForHosts "${sbHost} ${nodeHost}" 
@@ -53,11 +62,15 @@ if [ "${action}" = "list-devices" ]; then
 
 elif [ "${action}" = "install" ]; then
     test -z "${device}" && echo 'error: missing required parameter: -d [device]' 1>&2 && exit 1
+    test -z "${lxcFs}" && echo 'error: missing required parameter: -f [lxc-filesystem]' 1>&2 && exit 1
 
     installAccessForSshHost $nodeHost
     
     rsync -azve "ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no'" libfns.sh $nodeHost:/tmp/
-    ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no' $nodeHost "source /tmp/libfns.sh && prepareNode ${device}"
+    abortIfNonZero $? 'rsync libfns.sh failed'
+
+    ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no' $nodeHost "source /tmp/libfns.sh && prepareNode ${device} ${lxcFs}"
+    abortIfNonZero $? 'remote prepareNode() invocation'
 
 else
 	echo 'unrecognized action: ${action}' 1>&2 && exit 1
