+++ install/shipbuilder.sh
@@ -4,18 +4,17 @@ cd "$(dirname "$0")"
 
 source libfns.sh
 
-RESOURCES='server.sh setupBtrfs.sh loadBalancer.sh'
-
-while getopts “d:S:h” OPTION; do
+while getopts "d:f:S:h" OPTION; do
     case $OPTION in
         h)
-            echo "usage: $0 -S [shipbuilder-host] -d [server-btrfs-device] ACTION" 1>&2
+            echo "usage: $0 -S [shipbuilder-host] -d [server-dedicated-device] -f [lxc-filesystem] ACTION" 1>&2
             echo '' 1>&2
             echo 'This is the ShipBuilder installer program.' 1>&2
             echo '' 1>&2
-            echo '  ACTION                      Action to perform. Available actions are: install, list-devices'
-            echo '  -S [shipbuilder-host]       ShipBuilder Server SSH address (e.g. ubuntu@my.sb)' 1>&2
-            echo '  -d [sb-server-btrfs-device] Device to format with BTRFS (e.g. /dev/xvdc)' 1>&2
+            echo '  ACTION                       Action to perform. Available actions are: install, list-devices'
+            echo '  -S [shipbuilder-host]        ShipBuilder server user@hostname (flag can be omitted if auto-detected from env/SB_SSH_HOST)' 1>&2
+            echo '  -d [server-dedicated-device] Device to format with btrfs or zfs filesystem and use to store lxc containers (e.g. /dev/xvdc)' 1>&2
+            echo '  -f [lxc-filesystem]          LXC filesystem to use; "zfs" or "btrfs" (flag can be ommitted if auto-detected from env/LXC_FS)' 1>&2
             exit 1
             ;;
         S)
@@ -23,6 +22,9 @@ while getopts “d:S:h” OPTION; do
             ;;
         d)
             device=$OPTARG
+            ;;
+        f)
+            lxcFs=$OPTARG
             ;;
     esac
 done
@@ -33,9 +35,14 @@ shift $(($OPTIND - 1))
 action=$1
 
 test -z "${sbHost}" && autoDetectServer
+test -z "${lxcFs}" && autoDetectFilesystem
 
 test -z "${sbHost}" && echo 'error: missing required parameter: -S [shipbuilder-host]' 1>&2 && exit 1
-test -z "${action}" && echo 'error: missing required parameter: action' 1>&2 && exit 1
+#test -z "${action}" && echo 'error: missing required parameter: action' 1>&2 && exit 1
+if test -z "${action}"; then
+    echo 'info: action defaulted to: install'
+    action='install'
+fi
 
 
 verifySshAndSudoForHosts "${sbHost}"
@@ -52,16 +59,24 @@ if [ "${action}" = "list-devices" ]; then
 
 elif [ "${action}" = "install" ]; then
     test -z "${device}" && echo 'error: missing required parameter: -d [device]' 1>&2 && exit 1
+    test -z "${lxcFs}" && echo 'error: missing required parameter: -f [lxc-filesystem]' 1>&2 && exit 1
 
     installAccessForSshHost $sbHost
+    abortIfNonZero $? 'installAccessForSshHost() failed'
 
     rsync -azve "ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no'" libfns.sh $sbHost:/tmp/
-    ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no' $sbHost "source /tmp/libfns.sh && prepareServerPart1 ${sbHost} ${device}"
+    abortIfNonZero $? 'rsync libfns.sh failed'
+
+    ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no' $sbHost "source /tmp/libfns.sh && prepareServerPart1 ${sbHost} ${device} ${lxcFs}"
+    abortIfNonZero $? 'remote prepareServerPart1() invocation'
+
     mv ../env/SB_SSH_HOST{,.bak}
     echo "${sbHost}" > ../env/SB_SSH_HOST
     ../deploy.sh -f
     mv ../env/SB_SSH_HOST{.bak,}
-    ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no' $sbHost "source /tmp/libfns.sh && prepareServerPart2"
+
+    ssh -o 'BatchMode yes' -o 'StrictHostKeyChecking no' $sbHost "source /tmp/libfns.sh && prepareServerPart2 ${lxcFs}"
+    abortIfNonZero $? 'remote prepareServerPart2() invocation'
 
 else
     echo 'unrecognized action: ${action}' 1>&2 && exit 1
